+++ docs_md/changelog.md
@@ -138,3 +138,4 @@ Desc
 
 - Input component: rename field _selected_ to _is_selected_ (according to the docs)
 - Add EmmyLua annotations. See how to use it FAQ
+- Lang text now can be initialized without default locale id